+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-## [0.1.4] - 2020-11-19
+## Unreleased
 
+#### Fixed
+- Firebase tokens and messages forwarding.
+
+#### Changed
 - iOS SDK version updated to 4.6.0
 
 ## [0.1.3] - 2020-11-02